odk-frame-analyzer/pyproject.toml
[tool.poetry]
name = "odk-frame-analyzer-rewrite"
version = "0.1.0"
description = ""
authors = [
    "Stan Guldemond <s.guldemond@amsterdam.nl>",
    "Thomas Brockmeier <t.brockmeier@amsterdam.nl>"
]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python-headless = "^4.5.1"
pydantic = "^1.7.3"
aio-pika = "^6.7.1"
torch = "==1.6.0"
loguru = "^0.5.3"
typing = "^3.7.4"
imageio = "^2.9.0"
numpy = ">=1.18.5"
torchvision = "==0.7.0"
Cython = "^0.29.21"
matplotlib = ">=3.2.2"
Pillow = "^8.1.0"
PyYAML = ">=5.3"
scipy = ">=1.4.1"
tensorboard = ">=2.2"
tqdm = ">=4.41.0"
httpx = "^0.20.0"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"
pytest = "^6.2.1"
pytest-asyncio = "^0.14.0"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
